+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - 实现预设接收器。
   - com.dwarfeng.statistics.impl.handler.receiver.DoNothingReceiver。
   - com.dwarfeng.statistics.impl.handler.receiver.InjvmReceiver。
+  - com.dwarfeng.statistics.impl.handler.receiver.KafkaReceiver。
 
 - 实现预设调度器。
   - com.dwarfeng.statistics.impl.handler.dispatcher.DrainDispatcher。
   - com.dwarfeng.statistics.impl.handler.dispatcher.InjvmDispatcher。
+  - com.dwarfeng.statistics.impl.handler.dispatcher.KafkaDispatcher。
 
 - 实现预设驱动器。
   - com.dwarfeng.statistics.impl.handler.driver.CronDriverProvider。